% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/example.r
\name{example}
\alias{example}
\title{Example environmental time series data.}
\format{A data frame with 4220 rows and 3 columns:
\describe{
  \item{DateTime}{a date time object}
  \item{variable}{the observed variable}
  \item{value}{the value for that variable and time}
}}
\usage{
example(topic, package = NULL, lib.loc = NULL, character.only = FALSE,
  give.lines = FALSE, local = FALSE, echo = TRUE,
  verbose = getOption("verbose"), setRNG = FALSE,
  ask = getOption("example.ask"), prompt.prefix = abbreviate(topic, 6),
  run.dontrun = FALSE, run.donttest = interactive())
}
\description{
A tidy data set containing time-stamped observations of environmental data.
}

