% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mark.r
\name{mark}
\alias{mark}
\title{Marking time-series data}
\usage{
mark(data = NULL)
}
\arguments{
\item{data}{Your input data, should have a first column named 'DateTime' and the remaining columns as data.}
}
\description{
This function opens up the marking UI.
}
\examples{
\dontrun{
mark()
}
}

