% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/markr.r
\docType{package}
\name{markr}
\alias{markr}
\alias{markr-package}
\title{markr / Marking time series data in R}
\description{
markr provides a visual interface for rapid anomaly detection in time series data. It can be used as part of an analysis work flow or stand alone.
}
\details{
The package fits a machine learning algorithm (support vector machine) to automatically suggest anomalies. The user can keep those flags or mark new sections of the data and add flag names and comments. Marked data and flag metadata can be saved for future use. The fitted model objects can be saved locally for training future data sets as well.
}
\section{Marking data}{

 \code{\link{mark}}
}

